requirements.txt
streamlit
scikit-learn
pandas
numpy
mysql-connector-python
